+++ backup.py
@@ -18,9 +18,9 @@ class DBnew():
             c = conn.cursor()
 
             c.execute('''CREATE TABLE IF NOT EXISTS `user` (
-                `id` INTEGER NOT NULL PRIMARY KEY  AUTO_INCREMENT,
+                `id` INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT,
                 `email` VARCHAR(256) NOT NULL,
-                `email_verified` TINYINT(1) NOT NULL DEFAULT 0,
+                `email_verified` TINYINT NOT NULL DEFAULT 0,
                 `password` VARBINARY(128) NOT NULL,
                 `password_md5` VARBINARY(128) NOT NULL DEFAULT '',
                 `userkey` VARBINARY(128) NOT NULL,
@@ -32,17 +32,18 @@ class DBnew():
                 `cip` VARBINARY(16) NOT NULL,
                 `mip` VARBINARY(16) NOT NULL,
                 `aip` VARBINARY(16) NOT NULL,
-                `skey` VARBINARY(128) NOT NULL DEFAULT '',
-                `barkurl` VARBINARY(128) NOT NULL DEFAULT '',
-                `wxpusher` VARBINARY(128) NOT NULL DEFAULT '',
+                `skey` VARCHAR(128) NOT NULL DEFAULT '',
+                `barkurl` VARCHAR(128) NOT NULL DEFAULT '',
+                `wxpusher` VARCHAR(128) NOT NULL DEFAULT '',
                 `noticeflg` INT UNSIGNED NOT NULL DEFAULT 1,
-                `logtime`  VARBINARY(1024) NOT NULL DEFAULT '{"en":false,"time":"20:00:00","ts":0,"schanEn":false,"WXPEn":false}',
-                `status`  VARBINARY(1024) NOT NULL DEFAULT 'Enable',
-                `diypusher` VARBINARY(1024) NOT NULL DEFAULT '',
-                `qywx_token` VARBINARY(1024) NOT NULL DEFAULT '',
-                `tg_token` VARBINARY(1024) NOT NULL DEFAULT '',
-                `dingding_token` VARBINARY(1024) NOT NULL DEFAULT '',
-                `push_batch`  VARBINARY(1024) NOT NULL DEFAULT '{"sw":false,"time":0,"delta":86400}'
+                `logtime`  VARCHAR(1024) NOT NULL DEFAULT '{"en":false,"time":"20:00:00","ts":0,"schanEn":false,"WXPEn":false}',
+                `status`  VARCHAR(1024) NOT NULL DEFAULT 'Enable',
+                `diypusher` VARCHAR(1024) NOT NULL DEFAULT '',
+                `qywx_token` VARCHAR(1024) NOT NULL DEFAULT '',
+                `qywx_webhook` VARCHAR(1024) NOT NULL DEFAULT '',
+                `tg_token` VARCHAR(1024) NOT NULL DEFAULT '',
+                `dingding_token` VARCHAR(1024) NOT NULL DEFAULT '',
+                `push_batch` VARCHAR(1024) NOT NULL DEFAULT '{"sw":false,"time":0,"delta":86400}'
                 );
                 CREATE TABLE IF NOT EXISTS `tpl` (
                 `id` INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT,
@@ -131,10 +132,10 @@ class DBnew():
             # 获取数据库信息            
             userid = int(userid)
             user = maindb.db.user.get(id=userid, fields=('id', 'email', 'email_verified', 'password', 'password_md5', 'userkey', 'nickname', 'role', 'ctime', 'mtime', 'atime', 'cip',
-                                      'mip', 'aip', 'skey', 'barkurl', 'wxpusher', 'noticeflg', 'logtime', 'status', 'notepad', 'diypusher', 'qywx_token', 'tg_token', 'dingding_token', 'push_batch'))
+                                      'mip', 'aip', 'skey', 'barkurl', 'wxpusher', 'noticeflg', 'logtime', 'status', 'notepad', 'diypusher', 'qywx_token', 'tg_token', 'dingding_token', 'qywx_webhook', 'push_batch'))
             userkey = maindb.db.user.__getuserkey(user['env'])
             tpls = []
-            for tpl in maindb.db.tpl.list(fields=('id', 'userid', 'siteurl', 'sitename', 'banner', 'disabled', 'public', 'lock', 'fork', 'har', 'tpl', 'variables', 'interval', 'note', 'success_count', 'failed_count', 'last_success', 'ctime', 'mtime', 'atime', 'tplurl', 'updateable', '_groups'), limit=None):
+            for tpl in maindb.db.tpl.list(fields=('id', 'userid', 'siteurl', 'sitename', 'banner', 'disabled', 'public', 'lock', 'fork', 'har', 'tpl', 'variables', 'interval', 'note', 'success_count', 'failed_count', 'last_success', 'ctime', 'mtime', 'atime', 'tplurl', 'updateable', '_groups', 'init_env'), limit=None):
                 if tpl['userid'] == userid:
                     tpls.append(tpl)
             tasks = []